requirements.txt
requests==2.18.4
python-dotenv ==0.10.3
apscheduler == 3.6.3